+++ src/more/more.styled.ts
@@ -1,6 +1,7 @@
 import styled from 'styled-components/native'
 import * as R from 'ramda'
 import { TEXT_SIZE } from '../core/text/text.constants'
+import { COLOR } from '../core/colors/colors.constants'
 
 interface MoreContainerProps {
   topInset: number;
@@ -14,6 +15,7 @@ export const MoreContainer = styled.View<MoreContainerProps>`
 
 export const MoreDisplayText = styled.Text`
   font-size: ${TEXT_SIZE.S}px;
+  color: ${COLOR.BLACK}
 `
 
 export const MoreActivityIndicatorWrapper = styled.View`